requirements.txt
# Core dependencies for WooCommerce Order Analysis
pandas>=2.0.0
openpyxl>=3.1.0
woocommerce>=3.0.0
httpx>=0.24.0
requests>=2.31.0
urllib3>=2.0.0

# Database support
mysql-connector-python>=8.0.0

# Data visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Development and testing (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0